requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blis==0.7.11
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.5
cymem==2.0.11
dataclasses-json==0.6.7
defusedxml==0.7.1
distro==1.9.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
filelock==3.18.0
filetype==1.2.0
Flask==2.2.5
fonttools==4.56.0
fpdf2==2.8.2
frozenlist==1.5.0
fsspec==2025.3.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.165.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.21
langchain-community==0.3.20
langchain-core==0.3.48
langchain-text-splitters==0.3.7
langcodes==3.5.0
langsmith==0.1.147
language_data==1.3.0
llvmlite==0.44.0
lxml==5.3.1
marisa-trie==1.2.1
MarkupSafe==3.0.2
marshmallow==3.26.1
more-itertools==10.6.0
mpmath==1.3.0
multidict==6.2.0
murmurhash==1.0.12
mypy-extensions==1.0.0
networkx==3.4.2
numba==0.61.2
numpy==1.26.4
orjson==3.10.15
packaging==23.2
pillow==11.1.0
preshed==3.0.9
propcache==0.3.0
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydub==0.25.1
pyparsing==3.2.3
PyPDF2==3.0.1
python-docx==0.8.11
python-dotenv==1.0.0
python-magic==0.4.27
pymongo==4.13.0
dnspython==2.7.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
safetensors==0.5.3
six==1.17.0
smart-open==6.4.0
sniffio==1.3.1
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.39
srsly==2.5.1
sympy==1.13.1
tenacity==8.5.0
thinc==8.2.5
tiktoken==0.9.0
tokenizers==0.19.1
tqdm==4.67.1
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
wasabi==1.1.3
weasel==0.3.4
Werkzeug==2.2.3
yarl==1.18.3
